Move artifacts management in utils
